+++ CHANGELOG.md
@@ -12,6 +12,9 @@
   the new object.  
   [Samuel Giddins](https://github.com/segiddins)
 
+* Podfile objects are equatable.  
+  [Samuel Giddins](https://github.com/segiddins)
+
 
 ## 1.5.0 (2018-04-04)